requirements.txt
torch==1.10.1
transformers==4.16.2
sentence-transformers==2.1.0
pandas==1.1.2
faiss-cpu==1.6.1
numpy>=1.20.1
folium==0.2.1
networkx
pyvis
gensim
pymongo
python-dotenv
fastapi
pydantic
uvicorn